EOS-26572: review comments and adhoc changes
